Move open_with_perm from core.util to core.path
Move its tests, too.
