+++ pyanaconda/installation.py
@@ -28,6 +28,7 @@ from pyanaconda.modules.common.util import is_module_available
 from pyanaconda.progress import progress_message, progress_step, progress_complete, progress_init
 from pyanaconda import flags
 from pyanaconda.core import util
+from pyanaconda.core.path import open_with_perm
 from pyanaconda import timezone
 from pyanaconda import network
 from pyanaconda.core.i18n import N_
@@ -67,7 +68,7 @@ def _writeKS(ksdata):
     path = conf.target.system_root + "/root/anaconda-ks.cfg"
 
     # Make it so only root can read - could have passwords
-    with util.open_with_perm(path, "w", 0o600) as f:
+    with open_with_perm(path, "w", 0o600) as f:
         f.write("# Generated by Anaconda {}\n".format(util.get_anaconda_version_string()))
         f.write(str(ksdata))
